churn_project_production/requirements.txt
numpy==2.1.2
pandas==2.2.3
PyYAML==6.0.2
scikit_learn==1.5.2
scipy==1.14.1